new home html, css, js
nothing is formatted or in correct spot in layout. provided basic html,
css, and js.

photoshop folder holds all the pictures that ill be making for the
website, ideas holds the statically rendered designs.

not sure what the-1train_files/…..@N0x.jpg files are but I’m uploading
so nothing breaks without it